COB: implemented AimWeapon
same as with StartBuilding, heading and pitch are now calculated
in floats, and only converted to short TAANGs in CCobInstance.

+++ rts/Sim/Units/COB/LuaUnitScript.cpp
@@ -344,7 +344,7 @@ int CLuaUnitScript::QueryWeapon(int weaponNum)
 }
 
 
-void CLuaUnitScript::AimWeapon(int weaponNum, int heading, int pitch)
+void CLuaUnitScript::AimWeapon(int weaponNum, float heading, float pitch)
 {
 }
